+++ src/components/molecules/EarthEditor/LayerTreeViewItem/LayerActions/index.tsx
@@ -1,11 +1,12 @@
 import React from "react";
+import useFileInput from "use-file-input";
+import { useIntl } from "react-intl";
 
 import { styled } from "@reearth/theme";
-import useHooks from "./hooks";
 import HelpButton from "@reearth/components/atoms/HelpButton";
 import Icon from "@reearth/components/atoms/Icon";
-import { useIntl } from "react-intl";
-import { Format } from "../Layer";
+
+export type Format = "kml" | "czml" | "geojson" | "shape" | "reearth";
 
 export type Props = {
   rootLayerId?: string;
@@ -23,9 +24,25 @@ const LayerActions: React.FC<Props> = ({
   onLayerGroupCreate,
 }) => {
   const intl = useIntl();
-  const { importLayer } = useHooks({
-    onLayerImport,
-  });
+  const importLayer = useFileInput(
+    (files: FileList) => {
+      const file = files[0];
+      if (!file) return;
+
+      const extension = file.name.slice(file.name.lastIndexOf(".") + 1);
+      const format: Format | undefined = ["kml", "czml", "geojson"].includes(extension)
+        ? (extension as Format)
+        : extension == "json"
+        ? "reearth"
+        : ["zip", "shp"].includes(extension)
+        ? "shape"
+        : undefined;
+      if (!format) return;
+
+      onLayerImport?.(file, format);
+    },
+    { accept: ".kml,.czml,.geojson,.shp,.zip,.json" },
+  );
 
   return (
     <ActionWrapper
