+++ src/util/use-double-click.ts
@@ -1,18 +1,20 @@
 import { useCallback, useRef } from "react";
 
 const useDoubleClick = (
-  onClick: () => void,
+  onClick: (() => void) | undefined,
   onDoubleClick: (() => void) | undefined,
 ): [() => void, () => void] => {
   const t = useRef<NodeJS.Timeout>();
 
   const handleClick = useCallback(() => {
-    t.current && clearTimeout(t.current);
-    t.current = setTimeout(onClick, 200);
+    if (t.current) clearTimeout(t.current);
+    if (onClick) {
+      t.current = setTimeout(onClick, 200);
+    }
   }, [onClick]);
 
   const handleDoubleClick = useCallback(() => {
-    t.current && clearTimeout(t.current);
+    if (t.current) clearTimeout(t.current);
     onDoubleClick?.();
   }, [onDoubleClick]);
